// src/app/pages/wildlife-advice-page/models/pollinator-visible.model.ts
export interface GridResponse { 
    "Title": String;
    "LatinName": String;
    "CommonName": String;
    "Hardiness": String;
    "Soil": String;
    "SoilPH": String;
    "Shadiness": String;
    "Moisture": String;
    "Habit": String;
    "Height": String;
    "Growth": String;
    "Native": String;
    "Pathname": String;
    "Name": String;
    "Username": String;
    "Copyright": String;
    "Link": String;
}